Fix patch.
