+++ pkg/util/client/client.go
@@ -18,6 +18,7 @@ package client
 
 import (
 	"context"
+	"fmt"
 
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
