+++ pkg/workload/workload_test.go
@@ -1300,40 +1300,3 @@ func TestNeedsSecondPass(t *testing.T) {
 		})
 	}
 }
-
-// func TestNominatedClusterNamesRules(t *testing.T) {
-// 	clusterName := "cluster-1"
-// 	nominatedClusters := []string{"cluster-1", "cluster-2"}
-
-// 	// Rule: when setting clusterName it must be one of the nominatedClusterNames
-// 	ws := WorkloadStatus{
-// 		NominatedClusterNames: nominatedClusters,
-// 	}
-// 	ws.ClusterName = &clusterName
-// 	if ws.ClusterName != nil && len(ws.NominatedClusterNames) > 0 {
-// 		found := false
-// 		for _, n := range ws.NominatedClusterNames {
-// 			if n == *ws.ClusterName {
-// 				found = true
-// 				break
-// 			}
-// 		}
-// 		if !found {
-// 			t.Errorf("clusterName %q must be in nominatedClusterNames %v", *ws.ClusterName, ws.NominatedClusterNames)
-// 		}
-// 	}
-
-// 	// Rule: nominatedClusterNames must not be changed when clusterName is set
-// 	oldWs := ws
-// 	ws.NominatedClusterNames = []string{"cluster-3"}
-// 	if oldWs.ClusterName != nil && ws.ClusterName != nil && !equalStringSlices(oldWs.NominatedClusterNames, ws.NominatedClusterNames) {
-// 		t.Errorf("nominatedClusterNames must not be changed when clusterName is set")
-// 	}
-
-// 	// Rule: clusterName is immutable once set
-// 	oldWs.ClusterName = &clusterName
-// 	ws.ClusterName = ptrToString("cluster-2")
-// 	if oldWs.ClusterName != nil && ws.ClusterName != nil && *oldWs.ClusterName != *ws.ClusterName {
-// 		t.Errorf("clusterName is immutable once set")
-// 	}
-// }
